Wrong sprintf argument size.
git-svn-id: https://omniorb.svn.sourceforge.net/svnroot/omniorb/branches/4_1@5956 31ce16a7-c309-44a6-b7c4-d857baf782c7

+++ omniORBpy/modules/omnipy.cc
@@ -766,7 +766,7 @@ extern "C" {
       char* err = new char[80];
       sprintf(err, "Operation requires %d argument%s; %d given",
 	      arg_len, (arg_len == 1) ? "" : "s",
-	      PyTuple_GET_SIZE(op_args));
+	      (int)PyTuple_GET_SIZE(op_args));
 
       PyErr_SetString(PyExc_TypeError, err);
       delete [] err;
